implement _config.display, move remaning locale utilities; update tests
