+++ pandas/core/config_init.py
@@ -14,7 +14,6 @@ from pandas._config.config import (
     is_bool, is_callable, is_instance_factory, is_int, is_one_of_factory,
     is_text)
 
-from pandas.io.formats.console import detect_console_encoding
 from pandas.io.formats.terminal import is_terminal
 
 # compute
@@ -110,16 +109,6 @@ pc_nb_repr_h_doc = """
     pandas objects (if it is available).
 """
 
-pc_date_dayfirst_doc = """
-: boolean
-    When True, prints and parses dates with the day first, eg 20/01/2005
-"""
-
-pc_date_yearfirst_doc = """
-: boolean
-    When True, prints and parses dates with the year first, eg 2005/01/20
-"""
-
 pc_pprint_nest_depth = """
 : int
     Controls the number of nested levels to process when pretty-printing
@@ -129,13 +118,6 @@ pc_multi_sparse_doc = """
 : boolean
     "sparsify" MultiIndex display (don't display repeated
     elements in outer levels within groups)
-"""
-
-pc_encoding_doc = """
-: str/unicode
-    Defaults to the detected encoding of the console.
-    Specifies the encoding to be used for strings returned by to_string,
-    these are generally strings meant to be displayed on the console.
 """
 
 float_format_doc = """
@@ -331,16 +313,10 @@ with cf.config_prefix('display'):
                        validator=is_text)
     cf.register_option('notebook_repr_html', True, pc_nb_repr_h_doc,
                        validator=is_bool)
-    cf.register_option('date_dayfirst', False, pc_date_dayfirst_doc,
-                       validator=is_bool)
-    cf.register_option('date_yearfirst', False, pc_date_yearfirst_doc,
-                       validator=is_bool)
     cf.register_option('pprint_nest_depth', 3, pc_pprint_nest_depth,
                        validator=is_int)
     cf.register_option('multi_sparse', True, pc_multi_sparse_doc,
                        validator=is_bool)
-    cf.register_option('encoding', detect_console_encoding(), pc_encoding_doc,
-                       validator=is_text)
     cf.register_option('expand_frame_repr', True, pc_expand_repr_doc)
     cf.register_option('show_dimensions', 'truncate', pc_show_dimensions_doc,
                        validator=is_one_of_factory([True, False, 'truncate']))
